UniversalToJsonFormatter and RecordFormatter flatten
